Settings: server address and port can't restore to default
When click "Reset to default", get values of "SUPL_HOST" and
"SUPL_PORT" from "/etc/gps.conf", only if the value is not null,
will set values to setting provider.

If the value of "SUPL_HOST" and "SUPL_PORT" from "/etc/gps.conf"
is null, set the default value to setting provider.

CRs-Fixed: 682627

Change-Id: Ieb1f3714bf7fc09d911811604ded6557ecd11c88