Settings: fix automatically connect wrongly in WLAN
When auto connect is unchecked and in Simplied-Chinese, the ap
still be automatically connected.
The reason is that the key for auto connect is used wrongly in
WifiSettings, it's not consistent with the key used in
AdvancedWifiSettings, so it cannot get corresponding value from
settings provider.

Will update with right key in WifiSettings.

Change-Id: I273f25307c98be3e52045723ec2bae8cbb72ee68
CRs-Fixed: 750963