Settings: fix potential NPE when resetting profiles
When the user hits the positive button in the confirmation dialog to
reset profiles, we force a refresh of the profiles list. This can lead
to a NPE exception being thrown if it happens before the dialog actually
gets dismissed (and the ProfilesList fragment is actually attached)
since the reload list calls getPreferenceScreen(), which can be null,
since it is not attached.

After the dialog disappears, onResume() gets called in ProfilesList
anyways, so there's no need for a forced refresh.

Change-Id: I93de7115a69cc7545bf6a33de3a4abf5443658b9
Signed-off-by: Roman Birg <roman@cyngn.com>
(cherry picked from commit fa91465876198226cd770d59410b7a1c26f489f1)