Settings: merge ProfilesList into ProfilesSettings
ProfilesSettings previously had a viewpager with the profiles list, and
an app group list. Then we removed app groups, but the viewpager was
left over.

The main reason for getting rid of it now is to make sure that
refreshList() always works. Previously there was a small chance that
because it was in a viewpager, the preference may not have been attached
(at the time of the resetAll() dialog being shown and user pressing yes),
which caused the refresh method call to fail, since it needs access to
the preference screen, which is null at that time.

Change-Id: Ib9c3a27e3062660bd8a91998c20c15b4ffb4390b
Signed-off-by: Roman Birg <roman@cyngn.com>
(cherry picked from commit 371b12775631f492ba0084f682b2c0d080d745b4)