Settings: use WhisperPush v2 if it is installed in place of v1
This anticipates a forthcoming update to WhisperPush that will
have a renamed package. Modify PrivacySettings to check for the
presence of the update and prefer it to the original package
if present. The update is checked to see if it has a necessary
permission to avoid impersonators.

Change-Id: Ie9b5078c924a4795f877896709792f5ee1ac7207
(cherry picked from commit 95bf6ff1e1dfb967ed137fb0e6219330a3d807bf)