Settings: continue factory reset if user leaves activity
If the FRP dialog is up, and the user presses the home button, the FRP
task will finish, and then try to use the activity context to initiate
the wipe, and the fragment to retreive args, both of which is no longer
valid.

Cache all variables before the task begins, and use the application
context to complete the calls.

Ref: CYNGNOS-427

Change-Id: I5c8b286195a0b57fd879a1022d64ea548363b925
Signed-off-by: Roman Birg <roman@cyngn.com>
(cherry picked from commit b11123302979f1a5f54abf122f8664eb47ac37ea)