Settings : Save instance state for app group deletes
When deleting an app, inside of an app group in profiles,
if the user rotates device while confirmation is showing and
presses ok, it is a no-op. Reason is mPackageToDelete loses its
state when rotation is change.

Address this by persisting it in the state bundle.

issue-id: CYNGNOS-1119

Change-Id: I43dd9ccae7e16e92e5ac054f1b2ac0119cbfe60b
(cherry picked from commit a02a1d5b881d2df8fe3a5d7c5b3787846fdb3182)