Settings: regulatory: Wrap the layout in a ScrollView
This screen allows an optional image to be provided via overlay,
this image might be bigger than the screen height, and usually is
when the phone is in landscape. This cuts the image and only
shows what the screen allows.

This patch wraps it in a ScrollView to ensure that the user can
read the whole image.

Additionally, it changes the ImageView width to take the full width
of the parent, making centerCrop work as expected and centering the
image.

Change-Id: Ie8433767fe333cb4b019608208a45cd0a653641b
Ticket: HAM-138
(cherry picked from commit ff01cb15bc27eec94f7a757df1286bc03eff8865)