SimSettings: Disable entry for missing sim
Clicking on a missing SIM card will show incorrect information
because the user network preference is indexed by subscription id.

This means that when the SIM is not present we cannot infer which
subscription settings we are actually showing or changing so the
whole screen is effectively useless.

This patch disables configuration entries for invalid subscriptions.

Ticket: CYNGNOS-1794
Change-Id: Ic5125f66f4cdaafe06d03b3b20b80341525d6c87
(cherry picked from commit 389479f395acd02ff38610a5e677a923ab77744a)