Update textview styles for better theme compatability.
In the battery & notification lights preferences, the textview
attached to the custom color is locked to a color which is only
visible against light backgrounds. If a theme changes to a dark
theme, they are unreadable without modifying a core material
style.

Let's create a new style that points to a better material style
but also allows a themer to change it independently if needed.

Change-Id: I039f154282035c5123fddf8c20c18bdb3be90060
Ticket: CYNGNOS-1790
(cherry picked from commit 76ae8810a133b70099316314724bdd1e330fb21f)