SimDialogActivity: Check whether there is no default sub
On an out-of-the-box device, no data subscription will be
selected by default. If a single SIM is added, it will
become the default data subscription but if two sims are
added at the same time there will be no default.

In this case, this dialog will crash because it's trying
to get the subscription id from a null default. This patch
short-circuits the clause to always update if there is no
default.

Change-Id: Icd4dffaf918d43d85f807cd6a14f93abb118eca4
Ticket: CYNGNOS-1940
(cherry picked from commit 0eb5753e1c975967ef684514ab5f78f52d60b6db)