Re-expose hardcoded white text colors
In the original expose pass, all of the hardcoded white text colors
were given a generic color that they all pointed to. This is great,
but very inflexible for themes that might need to tweak just one
of those instances.

With this patch, we will create new text colors (and allow some
textviews to share it when it makes sense) and point all of those
back to the text_color_white so that existing themes are not
broken.

Change-Id: Iae850b7f75d319bd308f001058918e4466c70dca