MSIM: Fix erroneous defaults display in slot selection dialog
Initializing the index at zero was showing the first slot as the default
when none was chosen. Since that's not actually true, leave it outside
the possible range so that the radio selectors stay empty if that's the
case.

Addresses CYNGNOS-2212

Change-Id: I4a74f9c89bbf14bbe39e0d51733a1d5f92c5a1df
(cherry picked from commit 5da81a1e31c76ecca431076c9553be79095ca743)