Add new importance level for blocking lock screen notifications (2/2)
One feature that is still missing in the handling of notifications in
N is the possibility to block them app-wise for relevance/importance
reasons rather than for privacy reasons – i.e. also on the insecure
lock screen.

This patch plugs the hole by introducing a new importance level "very
low" between "min" and "low" in the power notification controls that
allows notifications on the status bar, but not on the lock screen.

It brings back most of the functionality of
Ib166db1b1673aeaea132c8eeb16c650d2f254a82 excluding the ability of
blocking only ongoing (persistent) notifications.

Change-Id: I01651c95a3fda68de6a2dd9e1acec5c9ea543f84
Signed-off-by: Marc K <morckx@gmail.com>